Time: 0 ms (100%), Space: 8.3 MB (49.35%) - LeetHub

+++ 0191-number-of-1-bits/0191-number-of-1-bits.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int hammingWeight(int n) {
-        long long i=0;
+     int i=0;
         while(n!=0){
             if((n&1)==1){
                 i++;
